requirements.txt
Pillow
scipy
numpy
matplotlib
